chore(ProspectAPI): Rename prospect-api-translation-lambda
