+++ .circleci/prospect-api.yml
@@ -46,7 +46,7 @@ workflows:
             aws-access-key-id: Dev_AWS_ACCESS_KEY
             aws-secret-access-key: Dev_AWS_SECRET_ACCESS_KEY
             aws-region: Dev_AWS_DEFAULT_REGION
-            scope: prospect-api-aws_lambda
+            scope: prospect-api-translation-aws_lambda
 
       - build_lambda:
             <<: *not_dev_main
@@ -56,7 +56,7 @@ workflows:
             aws-access-key-id: Dev_AWS_ACCESS_KEY
             aws-secret-access-key: Dev_AWS_SECRET_ACCESS_KEY
             aws-region: Dev_AWS_DEFAULT_REGION
-            scope: prospect-api-aws_lambda
+            scope: prospect-api-translation-aws_lambda
 
       - infrastructure:
           <<: *not_dev_main
@@ -122,7 +122,7 @@ workflows:
           aws-access-key-id: Dev_AWS_ACCESS_KEY
           aws-secret-access-key: Dev_AWS_SECRET_ACCESS_KEY
           aws-region: Dev_AWS_DEFAULT_REGION
-          scope: prospect-api-aws_lambda
+          scope: prospect-api-translation-aws_lambda
           s3-bucket: pocket-prospectapi-dev-sqs-translation
           requires:
             - prospect-api_infrastructure_apply_dev
@@ -243,7 +243,7 @@ workflows:
           aws-access-key-id: Prod_AWS_ACCESS_KEY
           aws-secret-access-key: Prod_AWS_SECRET_ACCESS_KEY
           aws-region: Prod_AWS_DEFAULT_REGION
-          scope: prospect-api-aws_lambda
+          scope: prospect-api-translation-aws_lambda
           s3-bucket: pocket-prospectapi-prod-sqs-translation
           requires:
             - prospect-api_infrastructure_apply_prod
