+++ infrastructure/prospect-api/src/translationSqsLambda.ts
@@ -1,28 +1,29 @@
 import { Construct } from 'constructs';
 import { config } from './config';
-import { PocketVPC } from '@pocket-tools/terraform-modules';
+import {
+  ApplicationDynamoDBTable,
+  PocketVPC,
+} from '@pocket-tools/terraform-modules';
 import { PocketSQSWithLambdaTarget } from '@pocket-tools/terraform-modules';
 import { LAMBDA_RUNTIMES } from '@pocket-tools/terraform-modules';
 import { DataAwsSsmParameter } from '@cdktf/provider-aws/lib/data-aws-ssm-parameter';
-import { DataAwsRegion } from '@cdktf/provider-aws/lib/data-aws-region';
-import { DataAwsCallerIdentity } from '@cdktf/provider-aws/lib/data-aws-caller-identity';
+import { PocketPagerDuty } from '@pocket-tools/terraform-modules';
 
-export class SqsLambdaBridge extends Construct {
+export class TranslationSqsLambda extends Construct {
   constructor(
     scope: Construct,
     private name: string,
-    dependencies: {
-      region: DataAwsRegion;
-      caller: DataAwsCallerIdentity;
-    },
+    prospectsTable: ApplicationDynamoDBTable,
+    pagerDuty?: PocketPagerDuty,
   ) {
     super(scope, name);
 
-    const { region, caller } = dependencies;
     const vpc = new PocketVPC(this, 'pocket-shared-vpc');
 
-    new PocketSQSWithLambdaTarget(this, 'bridge-sqs-lambda', {
-      name: `${config.prefix}-Sqs-Bridge`,
+    const { sentryDsn, gitSha } = this.getEnvVariableValues();
+
+    new PocketSQSWithLambdaTarget(this, 'translation-sqs-lambda', {
+      name: `${config.prefix}-Sqs-Translation`,
       // batch size is 1 so SQS doesn't get smart and try to combine them
       // (a combined message will mean a skipped candidate set from ML)
       batchSize: 1,
@@ -32,24 +33,31 @@ export class SqsLambdaBridge extends Construct {
         visibilityTimeoutSeconds: 300,
       },
       lambda: {
-        runtime: LAMBDA_RUNTIMES.NODEJS18,
+        runtime: LAMBDA_RUNTIMES.NODEJS14,
         handler: 'index.handler',
         timeout: 120,
         memorySizeInMb: 512,
         reservedConcurrencyLimit: 1,
         executionPolicyStatements: [
           {
-            actions: ['events:PutEvents'],
-            resources: [
-              `arn:aws:events:${region.name}:${caller.accountId}:event-bus/${config.envVars.eventBusName}`,
-            ],
             effect: 'Allow',
+            actions: [
+              'dynamodb:BatchWriteItem',
+              'dynamodb:PutItem',
+              'dynamodb:DescribeTable',
+              'dynamodb:UpdateItem',
+              'dynamodb:Query',
+            ],
+            resources: [
+              prospectsTable.dynamodb.arn,
+              `${prospectsTable.dynamodb.arn}/*`,
+            ],
           },
         ],
         environment: {
-          EVENT_BRIDGE_BUS_NAME: config.envVars.eventBusName,
-          SENTRY_DSN: this.getSentryDsn(),
-          GIT_SHA: this.getGitSha(),
+          PROSPECT_API_PROSPECTS_TABLE: prospectsTable.dynamodb.name,
+          SENTRY_DSN: sentryDsn,
+          GIT_SHA: gitSha,
           ENVIRONMENT:
             config.environment === 'Prod' ? 'production' : 'development',
         },
@@ -62,27 +70,32 @@ export class SqsLambdaBridge extends Construct {
           accountId: vpc.accountId,
         },
         alarms: {
-          // We don't configure alarms for this lambda, and
-          // instead rely on Sentry to alert on failure.
+          // TODO: set better alarm values
+          /*
+          errors: {
+            evaluationPeriods: 3,
+            period: 3600, // 1 hour
+            threshold: 20,
+            actions: config.isDev
+              ? []
+              : [pagerDuty!.snsNonCriticalAlarmTopic.arn],
+          },
+          */
         },
       },
       tags: config.tags,
     });
   }
 
-  private getSentryDsn() {
+  private getEnvVariableValues() {
     const sentryDsn = new DataAwsSsmParameter(this, 'sentry-dsn', {
       name: `/${config.name}/${config.environment}/SENTRY_DSN`,
     });
 
-    return sentryDsn.value;
-  }
-
-  private getGitSha() {
     const serviceHash = new DataAwsSsmParameter(this, 'service-hash', {
       name: `${config.circleCIPrefix}/SERVICE_HASH`,
     });
 
-    return serviceHash.value;
+    return { sentryDsn: sentryDsn.value, gitSha: serviceHash.value };
   }
 }
