+++ lambdas/prospect-api-translation-lambda/jest.config.js
