+++ lambdas/prospect-api-translation-lambda/src/config.ts
