+++ lambdas/prospect-api-translation-lambda/src/dynamodb/lib.integration.ts
