+++ lambdas/prospect-api-translation-lambda/src/dynamodb/lib.ts
