+++ lambdas/prospect-api-translation-lambda/src/index.ts
