+++ lambdas/prospect-api-translation-lambda/src/lib.spec.ts
