+++ lambdas/prospect-api-translation-lambda/src/lib.ts
