+++ lambdas/prospect-api-translation-lambda/src/types.ts
