+++ lambdas/prospect-api-translation-lambda/tsconfig.json
